+++ apps/greenhouse/src/components/admin/constants.ts
@@ -8,7 +8,7 @@ export const SELECTED_FILTER_PREFIX = "f_"
 export const NO_VALUE_DEFAULT = "--"
 
 export const SUPPORT_GROUP_LABEL = "greenhouse.sap/owned-by"
-export const EXPOSED_SERVICES_LABEL = "greenhouse.sap/plugin-exposed-services=true"
+export const EXPOSED_SERVICES_LABEL_SELECTOR = "greenhouse.sap/plugin-exposed-services=true"
 
 export const FILTER_IDS = {
   PLUGIN_PRESET_DEFINITION: "pluginPresetDefinition",
